+++ includes/publisher/js/src/components/RichText/RichTextEditorHeader.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+
+const RichTextEditorHeader = ({ modelSlug, field }) => {
+	return (
+		<>
+			<label
+				htmlFor={`atlas-content-modeler[${modelSlug}][${field.slug}]`}
+			>
+				{field.name}
+			</label>
+			{field?.description && (
+				<p className="help mb-4">{field.description}</p>
+			)}
+		</>
+	);
+};
+
+export default RichTextEditorHeader;
