+++ includes/publisher/js/src/components/RichText/index.js
@@ -1,0 +1,1 @@
+export { default } from "./RichTextEditor";
